minor changes and comments on how to fix randomness with seeding in Leo's parallel theards

+++ src/utils/randomSeed.hpp
@@ -1,8 +1,8 @@
 #ifndef RANDOMSEED_HPP_
 #define RANDOMSEED_HPP_
 
-void setSeed(unsigned int seed);
+void setSeed(unsigned long int seed);
 void setRandomSeed();
-unsigned int getRandomSeed();
+unsigned long int getRandomSeed();
 
 #endif /* RANDOMSEED_HPP_ */
